renamed loader modules to avoid confusion
